backend/requirements.txt
asgiref==3.9.2
certifi==2025.8.3
channels==4.3.1
charset-normalizer==3.4.3
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.10
pillow==11.3.0
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.5
sqlparse==0.5.3
stripe==12.5.1
typing_extensions==4.15.0
urllib3==2.5.0
APScheduler==3.10.4
channels-redis==4.2.0
django-redis==5.4.0
redis==5.0.8
sentry-sdk==2.17.0
